requirements.txt
numpy==1.26.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.2
scikit-learn==1.3.2
xgboost==1.7.6
statsmodels==0.14.1
shap==0.44.1
boruta==0.3
minisom==2.3.1
tensorflow==2.15.0
keras-tuner==1.4.6
pyfolio-reloaded==0.9.5
python-dateutil==2.9.0.post0
pytz==2024.1
tzdata==2024.1
